Make config like hass-virtual.
Make DOMAIN explicit.

Tidy switch code.
