+++ changelog
@@ -1,5 +1,9 @@
-0.7.0a2: make instance variables explicit
-0.7.0a1: support device/entity and standard integrations
+0.7.0a3:
+  tidy up code
+0.7.0a2:
+  make instance variables explicit
+0.7.0a1:
+  support device/entity and standard integrations
 
 0.6.3: restore switch state
 0.6.2: restore wrongly removed options
