+++ custom_components/momentary/config_flow.py
@@ -1,40 +1,38 @@
+"""
+This component provides support for a momentary switch.
+"""
 
 import logging
 import voluptuous as vol
+from typing import Any
 
 from homeassistant import config_entries, exceptions
 from homeassistant.const import CONF_PLATFORM
+from homeassistant.data_entry_flow import FlowResult
 
-from .const import (
-    ATTR_FILE_NAME,
-    ATTR_GROUP_NAME,
-    ATTR_SWITCHES,
-    DB_DEFAULT_SWITCHES_FILE,
-    DEFAULT_IMPORTED_NAME,
-    DOMAIN
-)
-from .db import Db
+from .const import *
+from .cfg import BlendedCfg, UpgradeCfg
 
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class MomentaryConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
+class MomentaryConfigFlow(config_entries.ConfigFlow, domain=COMPONENT_DOMAIN):
 
     VERSION = 1
 
     async def validate_input(self, user_input):
-        for group, values in self.hass.data.get(DOMAIN, {}).items():
+        for group, values in self.hass.data.get(COMPONENT_DOMAIN, {}).items():
             _LOGGER.debug(f"checking {group}")
             if group == user_input[ATTR_GROUP_NAME]:
                 raise GroupNameAlreadyUsed
             if values[ATTR_FILE_NAME] == user_input[ATTR_FILE_NAME]:
                 raise FileNameAlreadyUsed
         return {
-            "title": f"{user_input[ATTR_GROUP_NAME]} - {DOMAIN}"
+            "title": f"{user_input[ATTR_GROUP_NAME]} - {COMPONENT_DOMAIN}"
         }
 
-    async def async_step_user(self, user_input):
+    async def async_step_user(self, user_input: dict[str, Any] | None = None) -> FlowResult:
         _LOGGER.debug(f"step user {user_input}")
 
         errors = {}
@@ -42,16 +40,16 @@ class MomentaryConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
             try:
                 info = await self.validate_input(user_input)
 
-                db = Db(user_input[ATTR_GROUP_NAME], user_input[ATTR_FILE_NAME])
-                db.load()
+                cfg = BlendedCfg(user_input[ATTR_GROUP_NAME], user_input[ATTR_FILE_NAME])
+                cfg.load()
                 return self.async_create_entry(title=info["title"], data={
                     ATTR_GROUP_NAME: user_input[ATTR_GROUP_NAME],
                     ATTR_FILE_NAME: user_input[ATTR_FILE_NAME],
-                    ATTR_SWITCHES: list(db.switches.keys())
+                    ATTR_SWITCHES: list(cfg.switches.keys())
                 })
-            except GroupNameAlreadyUsed as e:
+            except GroupNameAlreadyUsed as _e:
                 errors["base"] = "group_name_used"
-            except FileNameAlreadyUsed as e:
+            except FileNameAlreadyUsed as _e:
                 errors["base"] = "file_name_used"
 
         else:
@@ -72,21 +70,20 @@ class MomentaryConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
     async def async_step_import(self, import_data):
         """Import momentary config from configuration.yaml."""
 
-        # Set up the new database.
-        _LOGGER.info("importing YAML switches into default group")
-        db = Db(DEFAULT_IMPORTED_NAME, DB_DEFAULT_SWITCHES_FILE)
-
         # Extract the momentary devices in the yaml file. The import function
         # converts it.
+        _LOGGER.info("importing YAML switches into default group")
+        cfg = UpgradeCfg(DEFAULT_IMPORTED_NAME, DB_DEFAULT_SWITCHES_FILE)
         for switch in import_data:
-            if switch[CONF_PLATFORM] == DOMAIN:
-                db.import_switch(switch)
+            if switch[CONF_PLATFORM] == COMPONENT_DOMAIN:
+                cfg.import_switch(switch)
+        cfg.save()
 
         # Store keys in momentary config.
-        return self.async_create_entry(title=f"{DEFAULT_IMPORTED_NAME} {DOMAIN}", data={
+        return self.async_create_entry(title=f"{DEFAULT_IMPORTED_NAME} {COMPONENT_DOMAIN}", data={
             ATTR_GROUP_NAME: DEFAULT_IMPORTED_NAME,
             ATTR_FILE_NAME: DB_DEFAULT_SWITCHES_FILE,
-            ATTR_SWITCHES: list(db.switches.keys())
+            ATTR_SWITCHES: list(cfg.switch_keys)
         })
